# [opt](serde)Optimize the filling of fixed values ​​into block columns without repeated deserialization.