--- a/be/src/vec/data_types/serde/data_type_decimal_serde.cpp
+++ b/be/src/vec/data_types/serde/data_type_decimal_serde.cpp
@@ -282,7 +282,8 @@ Status DataTypeDecimalSerDe<T>::deserialize_column_from_fixed_json(
         return st;
     }
 
-    DataTypeDecimalSerDe::insert_column_last_value_multiple_times(column, rows - column.size());
+    DataTypeDecimalSerDe::insert_column_last_value_multiple_times(column, rows - 1);
+    *num_deserialized = rows;
     return Status::OK();
 }