--- a/be/src/vec/data_types/serde/data_type_number_serde.cpp
+++ b/be/src/vec/data_types/serde/data_type_number_serde.cpp
@@ -233,7 +233,8 @@ Status DataTypeNumberSerDe<T>::deserialize_column_from_fixed_json(
         return st;
     }
 
-    DataTypeNumberSerDe::insert_column_last_value_multiple_times(column, rows - column.size());
+    DataTypeNumberSerDe::insert_column_last_value_multiple_times(column, rows - 1);
+    *num_deserialized = rows;
     return Status::OK();
 }