@hubgeter 2024-07-06
run buildall