@hubgeter 2024-07-07
run buildall