@hubgeter 2024-07-09
run buildall